synch

+++ java/src/main/java/com/techelevator/dao/JdbcCardsDao.java
@@ -2,6 +2,7 @@ package com.techelevator.dao;
 
 import com.techelevator.exception.DaoException;
 import com.techelevator.model.Cards;
+import com.techelevator.model.User;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.CannotGetJdbcConnectionException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -152,12 +153,16 @@ public class JdbcCardsDao implements CardsDao {
     }
 
     @Override
-    public boolean deleteCard(int cardId) {
+    public boolean deleteCard(int cardId, User user) {
+    Cards card = getCardById(cardId);
+    int deleteCardUserId = card.getUserId();
+    boolean isDeleted = false;
 
+    if(deleteCardUserId == user.getId()) {
 
         String sql1 = "DELETE from cards_tags where card_id = ?;";
         try {
-        jdbcTemplate.update(sql1, cardId);
+            jdbcTemplate.update(sql1, cardId);
         } catch (DataAccessException e) {
             throw new DaoException("Error deleting tags for this card");
         }
@@ -177,7 +182,9 @@ public class JdbcCardsDao implements CardsDao {
             throw new DaoException("Error deleting card with id: " + cardId, e);
 
         }
-        return true;
+        isDeleted = true;
+    }
+        return isDeleted;
 
     }
